+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Use `IndexMap` and `IndexSet` instead of the `std Hash*` equivalents to preserve ordering
 
 ## 0.27.1 - 2023-01-25
 ### Fixed